requirements.txt
albumentations==2.0.7
pytorch==2.1.0
pytorch-geometric==2.4.0
